+++ magda-gateway/src/createGenericProxyRouter.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import { Router } from "express";
-import _ from "lodash";
+import urijs from "urijs";
 import escapeStringRegexp from "escape-string-regexp";
 
 import buildJwt from "magda-typescript-common/src/session/buildJwt";
@@ -10,9 +10,14 @@ import Authenticator from "./Authenticator";
 import { TenantMode } from "./setupTenantMode";
 
 export type ProxyTarget = DetailedProxyTarget | string;
+export type MethodWithProxyTaget = {
+    method: string;
+    target?: string;
+};
+export type ProxyMethodType = string | MethodWithProxyTaget;
 export interface DetailedProxyTarget {
     to: string;
-    methods?: string[];
+    methods?: ProxyMethodType[];
     auth?: boolean;
     redirectTrailingSlash?: boolean;
     statusCheck?: boolean;
@@ -77,7 +82,7 @@ export default function createGenericProxyRouter(
     function proxyRoute(
         baseRoute: string,
         target: string,
-        verbs: string[] = ["all"],
+        verbs: ProxyMethodType[] = ["all"],
         auth = false,
         redirectTrailingSlash = false
     ) {
@@ -88,14 +93,31 @@ export default function createGenericProxyRouter(
             authenticator.applyToRoute(routeRouter);
         }
 
-        verbs.forEach((verb: string) =>
-            routeRouter[verb.toLowerCase()](
-                "*",
-                (req: express.Request, res: express.Response) => {
-                    proxy.web(req, res, { target });
+        verbs.forEach((verb: ProxyMethodType) => {
+            if (typeof verb === "string") {
+                routeRouter[verb.toLowerCase()](
+                    "*",
+                    (req: express.Request, res: express.Response) => {
+                        proxy.web(req, res, { target });
+                    }
+                );
+            } else {
+                const method: string = verb.method.toLowerCase();
+                if (!method) {
+                    throw new Error(
+                        "Invalid non-string proxy target method type"
+                    );
                 }
-            )
-        );
+                const runtimeTarget =
+                    typeof verb?.target === "string" ? verb.target : target;
+                routeRouter[method](
+                    "*",
+                    (req: express.Request, res: express.Response) => {
+                        proxy.web(req, res, { target: runtimeTarget });
+                    }
+                );
+            }
+        });
 
         if (redirectTrailingSlash) {
             // --- has to use RegEx as `req.originalUrl` will match both with & without trailing /
@@ -110,22 +132,40 @@ export default function createGenericProxyRouter(
         return routeRouter;
     }
 
-    _.forEach(options.routes, (value: ProxyTarget, key: string) => {
-        const target =
-            typeof value === "string"
-                ? getDefaultProxyTargetDefinition(value)
-                : value;
+    Object.keys(options.routes)
+        .sort((a, b) => {
+            // make sure route path has more path segment items will be installed first (i.e. higher priority when takes up requests)
+            const segmentLenA = urijs(a).segment().length;
+            const segmentLenB = urijs(b).segment().length;
+            if (segmentLenA < segmentLenB) {
+                return 1;
+            } else if (segmentLenA > segmentLenB) {
+                return -1;
+            } else if (a < b) {
+                return -1;
+            } else if (a > b) {
+                return 1;
+            } else {
+                return 0;
+            }
+        })
+        .map((key: string) => {
+            const value: ProxyTarget = options.routes[key];
+            const target =
+                typeof value === "string"
+                    ? getDefaultProxyTargetDefinition(value)
+                    : value;
 
-        const path = !key ? "/" : key[0] === "/" ? key : `/${key}`;
+            const path = !key ? "/" : key[0] === "/" ? key : `/${key}`;
 
-        proxyRoute(
-            path,
-            target.to,
-            target.methods,
-            !!target.auth,
-            target.redirectTrailingSlash
-        );
-    });
+            proxyRoute(
+                path,
+                target.to,
+                target.methods,
+                !!target.auth,
+                target.redirectTrailingSlash
+            );
+        });
 
     return router;
 }