+++ CHANGES.md
@@ -11,6 +11,7 @@
 -   Make PUT /auth/users/:userId API accept other user field rather then `isAdmin` field only
 -   Release `acs-cmd` & `org-tree` as separate NPM packages
 -   Add set admin function to `acs-cmd` commandline utility package
+-   #3024 Gateway will automatically proxy all GET requests received at `/api/v0/registry` to registry read-only nodes (excepts Hooks APIs)
 
 ## 0.0.58